Created ReservationController and built  phase 1 of the search function on /guests
When a request is made to /rsvp, the corresponding view is loaded by the controller. This page contains a form which allows a guest to enter in their email address. After the form has been submitted, a post request is made and the reserve() method gets executed. The method runs a query to fuzzy match the email address, if a match is found the guest is redirected to /thank-you. The $guest collection gets passed into the view and a confirmation page is executed.

Otherwise, if an email is not found the user will receive an error message.