Installed Livewire and Alpine. Began refactoring /rsvp to use Livewire.
Created 2 Livewire components & views. Guest/Rsvp.php is fired when the endpoint is initially requested. It prompts the user to enter in an email address. The component then queries the DB to return a guest collection with the matching email address. Set up validation ensure the email exists.

After the guest has been returned, the child component is loaded and the initial component is no longer rendered. GuestFound gets passed the $guest object and sends it to guest-found.blade.php to generate the RSVP form. Brought in Alpine to conditionally hide / show elements as needed.

Still need to connect the inputs to Livewire. Lots to do.