Finished homepage and configured Flickity to work with the Laravel filesystem
The filepath gets passed to the home view and returns an array of all of the image filenames in /images/memories. This is now dynamic. A picture can be dropped in manually, or perhaps guests can even have the ability to upload pictures of us from the wedding.

Worked on schedule page and fleshed out a dropdown component using Alpine.

Fixed a query error in GuestFound LW component.